federated_training/requirements.txt
# =======================
# CORE PYTORCH PACKAGES
# =======================
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# =======================
# FEDERATED LEARNING
# =======================
flwr==1.5.0

# =======================
# DATA PROCESSING
# =======================
numpy==1.24.3
pandas==2.1.1
Pillow==10.0.0
opencv-python==4.8.0.76
scikit-learn==1.3.2

# =======================
# UTILITIES
# =======================
matplotlib==3.7.3
PyYAML==6.0.1
tqdm==4.66.1
requests==2.31.0
protobuf==3.20.3
grpcio==1.57.0

# =======================
# OPTIONAL BUT SAFE
# =======================
seaborn==0.13.0
joblib==1.3.2
typing_extensions==4.8.0
